+++ src/pages/generate_report_page.cpp
@@ -40,7 +40,7 @@ std::string generate_report_page::handle_input(application &app,
     }
 
     m_date.month = static_cast<uint8_t>(std::stoi(match[1].str()));
-    m_date.year = static_cast<uint8_t>(std::stoi(match[2].str()));
+    m_date.year = static_cast<uint16_t>(std::stoi(match[2].str()));
 
     if (!m_date.is_valid()) {
       return "Invalid date range.\n";
